Use sessionId to find staff id when adding messages
Instead of getting the staff id from the client, we first validate the session
id and use it to find the user's information in the database.  When that's
done, we use the returned staff id for the message LastEditBy field.  This
structure allows for much better client/server separation and will be the
model for future database update/insert operations.

The MessagePanel used in PatientTab{Presenter,View} is temporary and will be
replaced with an mvp approved version.

Signed-off-by: Ian Hilt <ihilt@eastway.org>