Keep patient message
Instead of destroying the message when it is closed, keep it around.
This changes the behavior of the Add TreeItem.  Previously any
subsequent Add's from the tree were still born; now the Add brings back
the previously created message and any changes the user may have made.
However, it is destroyed when the patient tab is closed.

Signed-off-by: Ian Hilt <ihilt@eastway.org>