Implement the User model
This change integrates the User model into the rest of the code.  Using
UserImpl.getSessionId() instead of Cookies.getCookie("sessionId").

An inteface, User, was created to make the tracking of fields retrieved
from the database somewhat easier.

Since static fields are not serializable, a non-static class, UserData,
was created to house the data retrieved from the database. This data is
then copied to the UserImpl class, thereby making it available as a
class variable.

Signed-off-by: Ian Hilt <ihilt@eastway.org>